Update CondenseSkill to use new SemanticKernel version
Summary: This commit updates the CondenseSkill project to use the latest version of the SemanticKernel package, which introduces some breaking changes in the namespace and interface names. The commit also adds a SKFunctionInput attribute to the Condense method, to provide a description and a default value for the input parameter. Additionally, the commit updates the config.json file for the Condenser semantic function, to include the input parameter definition.