Update PRSkill to use latest SemanticKernel version and add input descriptions
Summary: This commit updates the PRSkill project to use the latest version of the SemanticKernel package, which requires some changes to the namespaces and types used. It also adds input descriptions to the SKFunction attributes and the config.json files of the semantic functions, to provide more information about the expected input format.