Add --ignore-space-change option to git show commands
Summary: This commit modifies the arguments of the git show commands used in the RunPullRequestDescription and RunPullRequestFeedback methods of the SKonsole program. The --ignore-space-change option is added to ignore changes in whitespace when comparing lines. This makes the output of the git show commands more readable and relevant, as it only shows the actual changes in the code logic and not the formatting.