👷 Update SemanticKernel dependencies and refactor code
This commit updates the SemanticKernel packages to the latest version
and refactors some of the code to use the new APIs. It also changes the
way the kernel is configured and adds a new HttpRetryConfig option. The
CondenseSkill and PullRequestSkill are also updated to use the new
kernel APIs and fix some minor issues.