📦 Update SemanticKernel packages
This commit updates the Microsoft.SemanticKernel and
Microsoft.SemanticKernel.Skills.Web packages to the latest preview
version (0.15.230609.2-preview) in SKonsole and the skills projects.
This ensures that the projects are using the most recent features and
bug fixes from the SemanticKernel library.

Additionally, this commit implements the ModelResult property in the
RedirectTextCompletionResult class, which was added to the
ITextCompletionResult interface in the new package version. This
property returns a ModelResult object that contains information about
the model that produced the completion.