Adding Generate PR Description workflow (#18)
This adds a PR comment-triggered workflow that uses the action [mkarle/skonsole-generate-pr-description](https://github.com/mkarle/skonsole-generate-pr-description).  That action uses [mkarle/pr-update](https://github.com/mkarle/pr-update) to update the PR.
To use, create a comment on a PR with:
`/sk generate-pr-description <suffix(default) | prefix | replace>`
Since it's triggered from a pull request comment, it only runs if it's in the default branch of the repo. [See this PR for an example of it working.](https://github.com/mkarle/skonsole/pull/4)

You'll need to add the following environment variables to the repo:
- variable: AZURE_OPENAI_CHAT_DEPLOYMENT_NAME
- secret: AZURE_OPENAI_API_ENDPOINT
- secret: AZURE_OPENAI_API_KEY