Enhancements, Refactoring, and Documentation Updates (#27)
* 🗑️ Remove EmailPlugin and EmailPluginEx classes

Removed the EmailPlugin and EmailPluginEx classes from the SKonsole
Plugins directory. This includes the deletion of SendEmail and
GetEmailAddress functions, as well as the Logger extension method.

* 🔄 Add cancellation token support to plugins

This commit adds support for CancellationToken in the CondensePlugin and
PRPlugin. It ensures that the token is checked during processing of
chunks and paragraphs, allowing for more responsive cancellation of
long-running tasks. This improves the overall performance and
responsiveness of the plugins.

* 🔄 Update logger instantiation in various classes

Replace specific class type with `this.GetType()` when creating loggers
in multiple classes, improving consistency and maintainability.

* 📦 Update GitPlugin with dynamic diff generation

This commit updates the GitPlugin to include dynamic diff generation
capabilities. The plugin now supports generating output based on a git
diff or git show file output for a given instruction. The update also
includes improvements to the GitDiffStaged function and adds a new
GitDiffDynamic function for more flexible diff filtering. The changes
made in this commit will enhance the overall functionality and
flexibility of the GitPlugin.

* 📝 Update README and improve SKonsole description

Update the README file to include a Contributing section and update the
license information. Improve the description of the SKonsole command in
Program.cs for better user understanding.

* 📝 Add validPlugins to contextVariables in StepwisePlannerCommand

This commit updates the StepwisePlannerCommand and LoadOptionSet methods to improve plugin handling and add
support for GitPlugin. The changes include:
- Adding validPlugins list to keep track of imported plugins.
- Modifying RunCreatePlan to pass validPlugins to RunChat.
- Updating RespondTo method to use validPlugins for filtering available functions.
- Implementing GitPlugin support with a new plan for processing 'git diff' output.

* 📦 Update SKonsole to v1.1.0 and reformat csproj

Update the SKonsole version from 1.0.6 to 1.1.0, and reformat the
csproj file for better readability. No changes in package
references or project references.

* 📦 Update SKonsole.csproj and add package metadata

Update SKonsole.csproj with additional package metadata, such as
owners, icon, repository type, tags, readme file, and license
expression. Also, include the icon and readme file in the package.