⚙️Major Refactor of GitPlugin, SuperFileIOPlugin, and PRPlugin
This commit includes significant changes to the GitPlugin, SuperFileIOPlugin, and PRPlugin classes, as well as updates to the associated configuration files. The main
focus of these changes is to update the code to work with the new Kernel updates.

In GitPlugin and SuperFileIOPlugin, the IKernel interface has been replaced with the Kernel class. The ITextCompletion service has been replaced with the
ITextGenerationService. The SKFunction attribute and SKContext class have been replaced with the KernelFunction attribute and KernelArguments class respectively.

In PRPlugin, the SKFunction attribute has been replaced with the KernelFunction attribute. The way functions are accessed has been changed from
`context.Functions.GetFunction` to `kernel.Plugins`. The `RunAsync` method has also been replaced with `InvokeAsync`.

The `config.json` files for the plugins have also been updated. The `completion` and `input` sections have been replaced with `execution_settings`.

These changes ensure that the plugins are compatible with the latest version of the Kernel and can utilize its new features and improvements.